Refactor mount() method binding feature and tests
