+++ tests/LifecycleHooksTest.php
@@ -2,10 +2,8 @@
 
 namespace Tests;
 
-use Livewire\Livewire;
 use Livewire\Component;
 use Livewire\LivewireManager;
-use Illuminate\Support\Facades\Route;
 
 class LifecycleHooksTest extends TestCase
 {
@@ -41,49 +39,6 @@ class LifecycleHooksTest extends TestCase
             'updatingFoo' => true,
             'updatedFoo' => true,
         ], $component->instance->lifecycles);
-    }
-
-    /** @test */
-    public function mount_hook_receives_route_model_bindings()
-    {
-        Livewire::component('foo', HasRouteModelBindingForMountHook::class);
-
-        Route::livewire('/test', 'foo');
-
-        $this->get('/test')->assertSee('output-from-method-binding');
-    }
-
-    /** @test */
-    public function mount_hook_method_receives_method_bindings()
-    {
-        Livewire::component('foo', HasRouteModelBindingForMountHook::class);
-
-        $output = view('render-component', [
-            'component' => 'foo',
-        ])->render();
-        
-        Route::get('/test', function() use ($output) {
-            return $output;
-        })->middleware('web');
-
-        $this->get('/test')->assertSee('output-from-method-binding');
-    }
-
-    /** @test */
-    public function mount_hook_method_receives_custom_bindings()
-    {
-        Livewire::component('foo', HasRouteModelBindingForMountHook::class);
-
-        Route::livewire('/test/{foo}', 'foo')->middleware('web');
-
-        Route::bind('foo', function ($value) {
-            $something = new ToBeBound;
-            $something->input = $value;
-
-            return $something;
-        });
-
-        $this->get('/test/should-show-up')->assertSee('should-show-up');
     }
 }
 
@@ -131,30 +86,5 @@ class ForLifecycleHooks extends Component
     public function render()
     {
         return app('view')->make('null-view');
-    }
-}
-
-class ToBeBound
-{
-    public $input = 'output-from-method-binding';
-
-    public function output()
-    {
-        return $this->input;
-    }
-}
-
-class HasRouteModelBindingForMountHook extends Component
-{
-    public $output;
-
-    public function mount(ToBeBound $toBeBound)
-    {
-        $this->output = $toBeBound->output();
-    }
-
-    public function render()
-    {
-        return app('view')->make('show-name', ['name' => $this->output]);
     }
 }
